+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Draw param to Cast functions
+
+### Changed
+- Improve Cast functions
 
 ## [1.0.1] - 2022-01-05
 ### Added